+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to the "selfie-grader" extension will be documented in this file.
 
+## 1.1.0
+- Extended README.md to include information about installation and WSL support
+- Added Option to hide assignment groups in `Grader Options`
+- Overhauled the `Grader Options` dropdown menu
+
 ## 1.0.1
 - Fixed excessive use of `\` in README.md